+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/GameChannel.java
@@ -32,7 +32,7 @@ public class GameChannel extends BaseEntity {
      */
     @TableId(type = IdType.AUTO)
     @ApiModelProperty(value = "自增主键")
-    private java.lang.Long id;
+    private java.lang.Integer id;
     /**
      * 渠道名称
      */
