+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/OpenServiceCampaignGiftDetail.java
@@ -27,19 +27,19 @@ public class OpenServiceCampaignGiftDetail extends BaseEntity {
      * id
      */
     @TableId(type = IdType.AUTO)
-    private java.lang.Integer id;
+    private java.lang.Long id;
 
     /**
      * 开服活动id
      */
     @Excel(name = "开服活动id", width = 15)
-    private java.lang.Integer campaignId;
+    private java.lang.Long campaignId;
 
     /**
      * open_service_campaign_type.id
      */
     @Excel(name = "open_service_campaign_type.id", width = 15)
-    private java.lang.Integer campaignTypeId;
+    private java.lang.Long campaignTypeId;
 
     /**
      * 活动名称
