+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/mapper/GameChannelMapper.java
@@ -27,7 +27,7 @@ public interface GameChannelMapper extends BaseMapper<GameChannel> {
             " s.`login_url`, s.`status`, s.`recommend`, s.`warning`, s.`min_version`, s.`max_version`" +
             " FROM `game_server` s LEFT JOIN `game_channel_server` c ON c.`server_id` = s.`id`" +
             " WHERE c.`channel_id` = #{channel_id}  AND c.`del_flag` = 0 ORDER BY c.`position`")
-    List<GameServerVO> getServerListChannelId(@Param("channel_id") Long channelId);
+    List<GameServerVO> getServerListChannelId(@Param("channel_id") Integer channelId);
 
     /**
      * 超找渠道名
